feat: set filter value
