+++ packages/shared/widget-plugin-external-events/src/hooks.ts
@@ -1,3 +1,3 @@
-export { useOnResetValueEvent } from "./hooks/useOnResetValueEvent";
+export { useOnResetValueEvent, useOnSetValueEvent } from "./hooks/useOnResetValueEvent";
 export { useOnResetFiltersEvent } from "./hooks/useOnResetFiltersEvent";
 export { useListenChannelEvents } from "./hooks/useListenChannelEvents";
