+++ packages/shared/widget-plugin-external-events/src/typings.ts
@@ -1,0 +1,8 @@
+export interface SetFilterValueArgs {
+    operators: string;
+    stringValue: string;
+    numberValue: number;
+    booleanValue: boolean;
+    dateTimeValue: Date;
+    dateTimeValue2: Date;
+}
