+++ packages/shared/widget-plugin-filtering/src/controls/input/typings.ts
@@ -49,12 +49,3 @@ export interface InputHookProps<TValue, TFilterEnum> {
     inputDisabled?: (filter: TFilterEnum) => boolean;
     inputType: "text" | "number";
 }
-
-export interface SetFilterValueArgs {
-    operators: string;
-    stringValue: string;
-    numberValue: number;
-    booleanValue: boolean;
-    dateTimeValue: Date;
-    dateTimeValue2: Date;
-}
